requirements.txt
Markdown==2.4
PyYAML==3.10
defusedxml==0.4.1
django-filter==0.7
django-oauth-plus==2.2.4
django-guardian==1.1.1
--------------------------
BeautifulSoup==3.2.1
boto==2.18.0
Django==1.6.2
Django-Select2==4.0.0
django-admin-honeypot==0.2.5
django-annoying==0.7.7
django-breadcrumbs==1.1.3-p1
django-cacheops==0.9.6
cache-tagging==0.7.7.0
django-celery==3.0.21
django-compressor==1.2
django-debug-toolbar==0.9.4
django-debug-toolbar-template-timings==0.5.1
django-filter==0.7
django-fixture-magic==0.0.4
django-generic-aggregation==0.3.2
django-grappelli==2.4.8
django-guardian==1.0.4
django-indexer==0.3.0
django-image-cropping==0.6.4
django-jfu==1.0.7
django-keyedcache==1.4-6
django-livesettings==1.4-12
django-markitup==1.0.0
django-memcache-status==1.1
django-modeltranslation==0.7.2
django-multiurl==1.0.1
django-olwidget==0.61.0
django-paging==0.2.4
django-positions==0.4.6
django-pure-pagination==0.2.1
djangorestframework==2.3.6
django-redis==3.4
django-redisboard==0.2.7
django-reversion==1.8.0
django-robots==0.8.1
django-rosetta==0.7.2
django-selectable==0.7.0
django-simple-captcha==0.4.0
django-smart-selects==1.0.9
django-taggit==0.11.2
django-taggit-autosuggest==0.2.1
django-templatetag-sugar==0.1
django-widget-tweaks==1.1.2
easy-thumbnails==1.2
johnny-cache==1.4
lxml==3.2.3
markdown==2.2.0
mongoengine==0.7.5
MySQL-python==1.2.3
Pillow==2.1.0
pymorphy==0.5.6
python-cdb==0.35
python-memcached==1.31
python-slugify==0.0.5
pymongo==2.2.1
pytils==0.2.3
redis==2.7.2
sorl-thumbnail==11.12
South==0.7.6
trans==1.4.2
translate==0.0.4
wsgiref==0.1.2
html2text==3.200.3
xhtml2pdf==0.0.5
sentry==5.4.5
raven==3.4.0
django-rq==0.5.1
rq==0.3.11
yandex.translate==0.3.4
grab==0.4.13
suds==0.4
uWSGI==2.0.1
reportlab==2.4